requirements.txt
glfw==2.6.2
pyglm==2.7.0
pyopengl==3.1.7
pyassimp==5.2.5
imgui==2.0.0
pillow==10.0.1
numpy==1.24.3